split StorageMiner into multiple files
